# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Medicology QBank API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateCurrentUser
      tags: [auth]
      summary: Update own profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /questions:
    get:
      operationId: getQuestions
      tags: [questions]
      summary: Get questions with filters
      security:
        - bearerAuth: []
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: system
          in: query
          schema:
            type: string
        - name: topic
          in: query
          schema:
            type: string
        - name: subtopic
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: university_tag
          in: query
          schema:
            type: string
        - name: exam_type
          in: query
          schema:
            type: string
            enum: [annual, modular]
        - name: tags
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: random
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionsResponse"
...
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a single question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
...
  /bookmarks:
    get:
      operationId: getBookmarks
      tags: [bookmarks]
      summary: Get user bookmarks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionsResponse"
    post:
      operationId: addBookmark
      tags: [bookmarks]
      summary: Add a bookmark
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bookmarks/{questionId}:
    delete:
      operationId: removeBookmark
      tags: [bookmarks]
      summary: Remove a bookmark
      security:
        - bearerAuth: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/questions:
    get:
      operationId: adminGetQuestions
      tags: [admin]
      summary: Admin - Get all questions
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionsResponse"
    post:
      operationId: adminCreateQuestion
      tags: [admin]
      summary: Admin - Create a question
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /admin/questions/{id}:
    put:
      operationId: adminUpdateQuestion
      tags: [admin]
      summary: Admin - Update a question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: adminDeleteQuestion
      tags: [admin]
      summary: Admin - Delete a question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Admin - Platform statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
...
  /admin/users/{id}:
    put:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Admin - Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Admin - Delete user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/reset-password:
    post:
      operationId: adminResetUserPassword
      tags: [admin]
      summary: Admin - Reset user password
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/flags:
    get:
      operationId: adminGetFlags
      tags: [admin]
      summary: Admin - Get all question flags
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFlagsResponse"
  /admin/flags/{id}:
    delete:
      operationId: adminDeleteFlag
      tags: [admin]
      summary: Admin - Delete a question flag
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        college:
          type: string
        university:
          type: string
        year:
          type: integer
          minimum: 1
          maximum: 6
      required: [name, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        college:
          type: string
        university:
          type: string
        year:
          type: integer
        isAdmin:
          type: boolean
        role:
          type: string
        rewardPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, college, year, isAdmin, role]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    QuestionOption:
      type: object
      properties:
        A:
          type: string
        B:
          type: string
        C:
          type: string
        D:
          type: string
        E:
          type: string
      required: [A, B, C, D]
    Question:
      type: object
      properties:
        id:
          type: integer
        questionText:
          type: string
        imageUrl:
          type: string
        options:
          $ref: "#/components/schemas/QuestionOption"
        correctAnswer:
          type: string
        explanation:
          type: string
        explanationImageUrl:
          type: string
        wrongAnswerExplanations:
          type: string
        references:
          type: string
        subject:
          type: string
        system:
          type: string
        topic:
          type: string
        subtopic:
          type: string
        universityTag:
          type: string
        examType:
          type: string
        difficulty:
          type: string
        tags:
          type: array
          items:
            type: string
        isBookmarked:
          type: boolean
        userNote:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, questionText, options, correctAnswer, explanation, subject, topic, difficulty]
    QuestionsResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [questions, total]
...
    BookmarkRequest:
      type: object
      properties:
        questionId:
          type: integer
      required: [questionId]
...
    CreateQuestionRequest:
      type: object
      properties:
        questionText:
          type: string
        options:
          $ref: "#/components/schemas/QuestionOption"
        correctAnswer:
          type: string
        imageUrl:
          type: string
        explanation:
          type: string
        explanationImageUrl:
          type: string
        wrongAnswerExplanations:
          type: string
        references:
          type: string
        subject:
          type: string
        system:
          type: string
        topic:
          type: string
        subtopic:
          type: string
        universityTag:
          type: string
        examType:
          type: string
          enum: [annual, modular]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: array
          items:
            type: string
      required: [questionText, options, correctAnswer, explanation, subject, topic, difficulty]
...
    AdminStats:
      type: object
      properties:
        totalQuestions:
          type: integer
        totalUsers:
          type: integer
        answersToday:
          type: integer
        pendingFlags:
          type: integer
        pendingErrata:
          type: integer
      required: [totalQuestions, totalUsers, answersToday, pendingFlags, pendingErrata]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        college:
          type: string
        university:
          type: string
        year:
          type: integer
        isAdmin:
          type: boolean
        role:
          type: string
        customPermissions:
          type: object
          additionalProperties:
            type: boolean
        rewardPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, isAdmin, role, rewardPoints, createdAt]
...
    AdminUpdateUserBody:
      type: object
      properties:
        isAdmin:
          type: boolean
        role:
          type: string
          enum: [user, editor, teacher, reviewer, admin, superadmin]
        customPermissions:
          type: object
          additionalProperties:
            type: boolean
        name:
          type: string
        email:
          type: string
        college:
          type: string
        university:
          type: string
        year:
          type: integer
...
    ResetPasswordBody:
      type: object
      properties:
        newPassword:
          type: string
      required: [newPassword]
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        college:
          type: string
        university:
          type: string
        year:
          type: integer
        currentPassword:
          type: string
        newPassword:
          type: string
    AdminFlag:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        questionId:
          type: integer
        userName:
          type: string
        questionText:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, questionId, createdAt]
    AdminFlagsResponse:
      type: object
      properties:
        flags:
          type: array
          items:
            $ref: "#/components/schemas/AdminFlag"
        total:
          type: integer
      required: [flags, total]
...